inital docker-compose setup.
